feat: enhance event and competitor management; add authorized roles and update scoring logic

+++ src/main/java/com/skylab/superapp/business/concretes/EventManager.java
@@ -2,17 +2,16 @@ package com.skylab.superapp.business.concretes;
 
 import com.skylab.superapp.business.abstracts.*;
 import com.skylab.superapp.core.constants.EventMessages;
-import com.skylab.superapp.core.exceptions.ImageAlreadyAddedException;
 import com.skylab.superapp.core.exceptions.ResourceNotFoundException;
 import com.skylab.superapp.core.mappers.EventMapper;
 import com.skylab.superapp.dataAccess.EventDao;
-import com.skylab.superapp.entities.Competition;
 import com.skylab.superapp.entities.DTOs.Event.CreateEventRequest;
 import com.skylab.superapp.entities.DTOs.Event.EventDto;
 import com.skylab.superapp.entities.DTOs.Event.UpdateEventRequest;
 import com.skylab.superapp.entities.Event;
 import com.skylab.superapp.entities.EventType;
 import com.skylab.superapp.entities.Image;
+import com.skylab.superapp.entities.Season;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Lazy;
@@ -20,8 +19,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 import java.util.UUID;
 
 @Service
@@ -30,37 +29,41 @@ public class EventManager implements EventService {
     private final EventDao eventDao;
     private final ImageService imageService;
     private final EventTypeService eventTypeService;
-    private final CompetitionService competitionService;
     private final Logger logger = LoggerFactory.getLogger(EventManager.class);
     private final EventMapper eventMapper;
+    private final UserService userService;
+    private final SeasonService seasonService;
+
+    private static final Set<String> PRIVILEGED_ROLES = Set.of("ADMIN", "YK", "DK");
 
     public EventManager(EventDao eventDao, @Lazy CompetitorService competitorService, @Lazy ImageService imageService,
-                        @Lazy EventTypeService eventTypeService,@Lazy CompetitionService competitionService, EventMapper eventMapper) {
+                        @Lazy EventTypeService eventTypeService, EventMapper eventMapper, UserService userService, SeasonService seasonService) {
         this.eventDao = eventDao;
         this.imageService = imageService;
         this.eventTypeService = eventTypeService;
-        this.competitionService = competitionService;
         this.eventMapper = eventMapper;
+        this.userService = userService;
+        this.seasonService = seasonService;
     }
 
     @Override
+    @Transactional
     public EventDto addEvent(CreateEventRequest createEventRequest, MultipartFile coverImageFile) {
         logger.info("Adding new event: {}", createEventRequest.getName());
-        EventType eventType = null;
-        if (createEventRequest.getEventTypeId() != null) {
-            eventType = eventTypeService.getEventTypeEntityById(createEventRequest.getEventTypeId());
-        }
 
-        Competition competition = null;
-        if (createEventRequest.getCompetitionId() != null) {
-            competition = competitionService.getCompetitionEntityById(createEventRequest.getCompetitionId());
+        EventType eventType = eventTypeService.getEventTypeEntityById(createEventRequest.getEventTypeId());
+
+        checkUserAuthorization(eventType);
+
+        Season season = null;
+        if(createEventRequest.getSeasonId() != null) {
+            season = seasonService.getSeasonEntityById(createEventRequest.getSeasonId());
         }
 
         Image savedCoverImage = null;
         if (coverImageFile != null && !coverImageFile.isEmpty()) {
             savedCoverImage = imageService.uploadImage(coverImageFile);
         }
-
 
 
         Event event = Event.builder()
@@ -73,8 +76,10 @@ public class EventManager implements EventService {
                 .active(createEventRequest.isActive())
                 .linkedin(createEventRequest.getLinkedin())
                 .location(createEventRequest.getLocation())
-                .competition(competition)
                 .coverImage(savedCoverImage)
+                .season(season)
+                .isRanked(createEventRequest.isRanked())
+                .prizeInfo(createEventRequest.getPrizeInfo())
                 .build();
 
         logger.info("Event created: {}", event.getName());
@@ -82,43 +87,48 @@ public class EventManager implements EventService {
     }
 
     @Override
+    @Transactional
     public void deleteEvent(UUID id) {
-        var event = getEventEntityById(id);
+        logger.info("Deleting event with id: {}", id);
+        Event event = getEventEntityById(id);
+
+        checkUserAuthorization(event.getType());
+
         eventDao.delete(event);
+
+        logger.info("Event deleted with id: {}", id);
     }
 
     @Override
+    @Transactional
     public EventDto updateEvent(UUID id, UpdateEventRequest updateEventRequest) {
+        logger.info("Updating event id: {}", id);
         var event = getEventEntityById(id);
-        logger.info("Updating event id: {}", event.getId());
 
-        event.setName(updateEventRequest.getName() == null
-                ? event.getName()
-                : updateEventRequest.getName());
-        event.setDescription(updateEventRequest.getDescription() == null
-                ? event.getDescription()
-                : updateEventRequest.getDescription());
-        event.setType(updateEventRequest.getType() == null
-                ? event.getType()
-                : eventTypeService.getEventTypeEntityByName(updateEventRequest.getType()));
-        event.setFormUrl(updateEventRequest.getFormUrl() == null
-                ? event.getFormUrl()
-                : updateEventRequest.getFormUrl());
-        event.setStartDate(updateEventRequest.getStartDate() == null
-                ? event.getStartDate()
-                : updateEventRequest.getStartDate());
-        event.setEndDate(updateEventRequest.getEndDate() == null
-                ? event.getEndDate()
-                : updateEventRequest.getEndDate());
-        event.setLinkedin(updateEventRequest.getLinkedin() == null
-                ? event.getLinkedin()
-                : updateEventRequest.getLinkedin());
+        checkUserAuthorization(event.getType());
+
+        if (updateEventRequest.getName() != null) event.setName(updateEventRequest.getName());
+        if (updateEventRequest.getDescription() != null) event.setDescription(updateEventRequest.getDescription());
+        if (updateEventRequest.getFormUrl() != null) event.setFormUrl(updateEventRequest.getFormUrl());
+        if (updateEventRequest.getStartDate() != null) event.setStartDate(updateEventRequest.getStartDate());
+        if (updateEventRequest.getEndDate() != null) event.setEndDate(updateEventRequest.getEndDate());
+        if (updateEventRequest.getLinkedin() != null) event.setLinkedin(updateEventRequest.getLinkedin());
+        if (updateEventRequest.getLocation() != null) event.setLocation(updateEventRequest.getLocation());
+
+        if (updateEventRequest.getPrizeInfo() != null) event.setPrizeInfo(updateEventRequest.getPrizeInfo());
+        event.setRanked(updateEventRequest.isRanked());
         event.setActive(updateEventRequest.isActive());
-        event.setCompetition(updateEventRequest.getCompetitionId() == null
-                ? event.getCompetition()
-                : competitionService.getCompetitionEntityById(updateEventRequest.getCompetitionId()));
 
-        logger.info("Event updated id: {}", event.getId());
+        if (updateEventRequest.getTypeId() != null) {
+            EventType newType = eventTypeService.getEventTypeEntityById(updateEventRequest.getTypeId());
+            event.setType(newType);
+        }
+
+        if (updateEventRequest.getSeasonId() != null) {
+            Season newSeason = seasonService.getSeasonEntityById(updateEventRequest.getSeasonId());
+            event.setSeason(newSeason);
+        }
+
         return eventMapper.toDto(eventDao.save(event));
     }
 
@@ -126,21 +136,21 @@ public class EventManager implements EventService {
     @Override
     public List<EventDto> getAllEventsByEventType(EventType eventType, boolean includeEventType, boolean includeSession,
                                                   boolean includeCompetitors, boolean includeImages,
-                                                  boolean includeSeason, boolean includeCompetition) {
+                                                  boolean includeSeason) {
         var eventTypeResult = eventTypeService.getEventTypeEntityById(eventType.getId());
 
         var list = eventDao.findAllByType(eventTypeResult);
         return eventMapper.toDtoList(list, includeEventType, includeSession,
-                includeCompetitors, includeImages, includeSeason, includeCompetition);
+                includeCompetitors, includeImages, includeSeason);
     }
 
     @Override
     public EventDto getEventById(UUID id, boolean includeEventType, boolean includeSession,
                                  boolean includeCompetitors, boolean includeImages,
-                                 boolean includeSeason, boolean includeCompetition) {
+                                 boolean includeSeason) {
 
         return eventMapper.toDto(getEventEntityById(id), includeEventType, includeSession,
-                includeCompetitors, includeImages, includeSeason, includeCompetition);
+                includeCompetitors, includeImages, includeSeason);
     }
 
 
@@ -148,6 +158,10 @@ public class EventManager implements EventService {
     @Transactional
     public void addImagesToEvent(UUID eventId, List<UUID> imageIds) {
         var event = getEventEntityById(eventId);
+
+        checkUserAuthorization(event.getType());
+
+
         var images = imageService.getImagesByIds(imageIds);
 
 
@@ -162,6 +176,8 @@ public class EventManager implements EventService {
     @Transactional
     public void removeImagesFromEvent(UUID eventId, List<UUID> imageIds) {
         Event event = getEventEntityById(eventId);
+
+        checkUserAuthorization(event.getType());
 
         List<Image> images = imageService.getImagesByIds(imageIds);
 
@@ -178,57 +194,71 @@ public class EventManager implements EventService {
     @Override
     public List<EventDto> getAllFutureEventsByEventType(String eventType, boolean includeEventType, boolean includeSession,
                                                         boolean includeCompetitors, boolean includeImages,
-                                                        boolean includeSeason, boolean includeCompetition) {
+                                                        boolean includeSeason) {
         var eventTypeResult = eventTypeService.getEventTypeEntityByName(eventType);
 
         var events = eventDao.findAllByType(eventTypeResult);
 
         return eventMapper.toDtoList(events, includeEventType, includeSession,
-                includeCompetitors, includeImages, includeSeason, includeCompetition);
+                includeCompetitors, includeImages, includeSeason);
     }
 
     @Override
     public List<EventDto> getAllEvents(boolean includeEventType, boolean includeSession,
                                          boolean includeCompetitors, boolean includeImages,
-                                         boolean includeSeason, boolean includeCompetition) {
+                                         boolean includeSeason) {
 
 
         var events = eventDao.findAll();
 
 
         return eventMapper.toDtoList(events, includeEventType, includeSession,
-                includeCompetitors, includeImages, includeSeason, includeCompetition);
+                includeCompetitors, includeImages, includeSeason);
     }
 
     @Override
     public List<EventDto> getAllEventsByEventTypeName(String eventTypeName, boolean includeEventType,
                                                       boolean includeSession, boolean includeCompetitors,
-                                                      boolean includeImages, boolean includeSeason,
-                                                      boolean includeCompetition) {
+                                                      boolean includeImages, boolean includeSeason) {
 
 
         var eventType = eventTypeService.getEventTypeEntityByName(eventTypeName);
         var events = eventDao.findAllByType(eventType);
 
         return eventMapper.toDtoList(events, includeEventType, includeSession,
-                includeCompetitors, includeImages, includeSeason, includeCompetition);
+                includeCompetitors, includeImages, includeSeason);
     }
 
     @Override
     public List<EventDto> getAllEventByIsActive(boolean isActive, boolean includeEventType, boolean includeSession,
                                                 boolean includeCompetitors, boolean includeImages,
-                                                boolean includeSeason, boolean includeCompetition) {
+                                                boolean includeSeason) {
 
         var events = eventDao.findAllByActive(isActive);
 
         return eventMapper.toDtoList(events, includeEventType, includeSession,
-                includeCompetitors, includeImages, includeSeason, includeCompetition);
+                includeCompetitors, includeImages, includeSeason);
     }
 
 
     @Override
     public Event getEventEntityById(UUID id) {
         return eventDao.findById(id).orElseThrow(()-> new ResourceNotFoundException(EventMessages.EVENT_NOT_FOUND));
+    }
+
+    private void checkUserAuthorization(EventType eventType) {
+        boolean authorizedUser = userService.getAuthenticatedUser()
+                .getRoles()
+                .stream()
+                .anyMatch(role ->
+                        PRIVILEGED_ROLES.contains(role) ||
+                                eventType.getAuthorizedRoles().contains(role)
+                );
+
+        if (!authorizedUser) {
+            logger.error("User not authorized for event type: {}", eventType.getName());
+            throw new ResourceNotFoundException(EventMessages.USER_NOT_AUTHORIZED_FOR_EVENT_TYPE);
+        }
     }
 
 }
